# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/race-history": {
            "get": {
                "description": "Get race history entries, optionally filtered by ID or type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Race"
                ],
                "summary": "Get race history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Race ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Race type (season, oneoff)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RaceHistory"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "Archive a completed race to history with results",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Race"
                ],
                "summary": "Save race to history",
                "parameters": [
                    {
                        "description": "Race history data with results",
                        "name": "race",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "Delete a race history entry by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Race"
                ],
                "summary": "Delete race from history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Race ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/race-info": {
            "get": {
                "description": "Get the current race information (country, track, laps)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Race"
                ],
                "summary": "Get current race info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RaceInfo"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "Update the current race information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Race"
                ],
                "summary": "Update race info",
                "parameters": [
                    {
                        "description": "Race info",
                        "name": "raceInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RaceInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.RaceHistory": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "race_date": {
                    "type": "string"
                },
                "race_type": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RaceResult"
                    }
                },
                "total_laps": {
                    "type": "integer"
                },
                "track": {
                    "type": "string"
                },
                "track_id": {
                    "type": "string"
                }
            }
        },
        "models.RaceInfo": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "laps": {
                    "type": "integer"
                },
                "next_race_date": {
                    "type": "string"
                },
                "track": {
                    "type": "string"
                },
                "track_id": {
                    "type": "string"
                }
            }
        },
        "models.RaceResult": {
            "type": "object",
            "properties": {
                "fastest_lap": {
                    "type": "boolean"
                },
                "finished": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                },
                "race_id": {
                    "type": "integer"
                },
                "racer_id": {
                    "type": "integer"
                },
                "racer_name": {
                    "type": "string"
                }
            }
        },
...
    }
}